+++ CHANGES.md
@@ -55,6 +55,7 @@ Changes:
   implementation can be found under `video.testsrc.ffmpeg`.
 - Images can now generate blank audio if needed, no need to add
   `mux_audio(audio=blank(),image)` anymore (#2230).
+- Removed deprecated `timeout` argument in `http.*` operators.
 
 2.0.2 (28-12-2021)
 =====